Colmap 3D Reconstruct/requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
packaging==25.0
pathlib==1.0.1
pillow==11.2.1
pycolmap==3.11.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
